requirements.txt
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
openai
python-dotenv
arize-phoenix
arize-phoenix-otel
openinference-instrumentation-openai
